fix(win): register CLI flags by moving the compile driver into codonc
On Windows, LLVM is statically linked into both codon.exe and codonc.dll, so
each owns a separate `llvm::cl` GlobalParser registry. The flag globals
(options.cpp, numpy.cpp's npfuse-*) live in codonc, but main.cpp parsed in the
exe's registry -> every codon-specific flag (-release, --numerics, -fast-math,
-disable-exceptions, npfuse-*) was rejected as "Unknown command line argument".
The npfuse-* flags are read inside codonc, so moving flags to the exe could not
fix it -- the registry had to be unified.

Move the whole compile driver (run/build/doc/jit + processSource + every
ParseCommandLineOptions call) out of main.cpp into a new codon/app/cli.cpp
compiled into codonc, exported as `codon::cliMain()` (codon/app/cli.h).
main.cpp is now a thin shim: it handles `jupyter` (depends on codon_jupyter,
not codonc flags) and forwards everything else. Not _WIN32-guarded -- a uniform
app/lib boundary change, harmless on Linux/macOS (single registry there
regardless).
